test: share external media e2e helpers
